Simplify project structure by removing codegen-test module
Integrated test functionality directly into the root project for better
maintainability. This reduces complexity and makes testing more straightforward.

- Removed separate codegen-test module and all its files
- Moved test configuration to root build.gradle.kts
- Integrated template files to root /template directory
- Updated CI workflow to reflect new structure
- Added PostgreSQL support via Testcontainers for testing
- Updated .gitignore for generated code paths
