+++ .github/workflows/ci.yml
@@ -19,22 +19,21 @@ jobs:
           distribution: 'zulu'
           java-version: 21
 
+      - name: Grant execute permission for codegen/gradlew
+        working-directory: ./codegen
+        run: chmod +x gradlew
+
+      - name: Build
+        working-directory: ./codegen
+        run: ./gradlew build
+
       - name: Grant execute permission for ./gradlew
         run: chmod +x gradlew
 
-      - name: Grant execute permission for codegen-test/gradlew
-        working-directory: ./codegen-test
-        run: chmod +x gradlew
-
-      - name: Assemble and Check
-        run: ./gradlew assemble check
-
       - name: Test plugin (generate Java code)
-        working-directory: ./codegen-test
         run: ./gradlew domaCodeGenJavaAll build
 
       - name: Test plugin (generate Kotlin code)
-        working-directory: ./codegen-test
         run: ./gradlew domaCodeGenKotlinAll build
 
       - name: Set version
@@ -46,4 +45,5 @@ jobs:
 
       - name: Publish plugin
         if: github.event_name == 'push' && endsWith(steps.set-version.outputs.version, 'SNAPSHOT') == false
+        working-directory: ./codegen
         run: ./gradlew publishPlugins -Pgradle.publish.key=${{ secrets.GRADLE_PUBLISH_KEY }} -Pgradle.publish.secret=${{ secrets.GRADLE_PUBLISH_SECRET }}
