+++ .gitignore
@@ -9,3 +9,11 @@
 /.idea/
 .factorypath
 /.claude/
+/data/
+/.apt_generated/
+
+/src/main/java/codegen
+/src/main/kotlin/codegen
+/src/main/resources/META-INF/codegen
+/src/test/java/codegen
+/src/test/kotlin/codegen
