--- codegen-test/.gitignore
@@ -1,12 +1,0 @@
-/.classpath
-/.gradle/
-/.project
-/.settings/
-/bin/
-/build/
-/src/
-/target/
-/data/
-/.metadata
-/.idea/
-/.apt_generated/
