--- codegen-test/docker-compose.yml
@@ -1,19 +1,0 @@
-# Use root/example as user/password credentials
-version: '3.1'
-
-services:
-
-  db:
-    image: mysql
-    command: --default-authentication-plugin=mysql_native_password
-    restart: always
-    environment:
-      MYSQL_ROOT_PASSWORD: example
-    ports:
-      - 3306:3306
-
-  adminer:
-    image: adminer
-    restart: always
-    ports:
-      - 8080:8080
