--- codegen-test/gradle.properties
@@ -1,2 +1,0 @@
-domaVersion=3.9.1
-h2Version=2.3.232
