--- codegen-test/settings.gradle.kts
@@ -1,5 +1,0 @@
-rootProject.name = "codegen-test"
-
-pluginManagement {
-    includeBuild("../codegen")
-}
