+++ gradle.properties
@@ -1,3 +1,4 @@
 group=org.domaframework.doma
 version=3.2.1-SNAPSHOT
 release.useAutomaticVersion=true
+domaVersion=3.9.1
